clean up old Policies
