+++ contracts/src/policy/Policy1.sol
@@ -4,21 +4,22 @@
 pragma solidity =0.8.6;
 
 import {SafeMath} from "@openzeppelin/contracts/utils/math/SafeMath.sol";
-import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
+import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {UsingRegistry} from "contracts/src/common/registry/UsingRegistry.sol";
 import {IPolicy} from "contracts/interface/IPolicy.sol";
+import {Curve} from "contracts/src/common/libs/Curve.sol";
+import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 
-/**
- * DIP1 is a contract that simply changed TheFirstPolicy to DIP numbering.
- */
-contract DIP1 is IPolicy, UsingRegistry {
+contract Policy1 is IPolicy, Ownable, Curve, UsingRegistry {
 	using SafeMath for uint256;
 	uint256 public override marketVotingBlocks = 525600;
 	uint256 public override policyVotingBlocks = 525600;
+	address private treasuryAddress;
+	address private capSetterAddress;
 
 	uint256 private constant basis = 10000000000000000000000000;
 	uint256 private constant power_basis = 10000000000;
-	uint256 private constant mint_per_block_and_aseet = 250000000000000;
+	uint256 private constant mint_per_block_and_aseet = 132000000000000;
 
 	constructor(address _registry) UsingRegistry(_registry) {}
 
@@ -29,29 +30,15 @@ contract DIP1 is IPolicy, UsingRegistry {
 		override
 		returns (uint256)
 	{
-		uint256 max = _assets.mul(mint_per_block_and_aseet);
-		uint256 t = ERC20(registry().registries("Dev")).totalSupply();
-		uint256 s = (_lockups.mul(basis)).div(t);
-		uint256 _d = basis.sub(s);
-		uint256 _p = (
-			(power_basis.mul(12)).sub(s.div((basis.div((power_basis.mul(10))))))
-		).div(2);
-		uint256 p = _p.div(power_basis);
-		uint256 rp = p.add(1);
-		uint256 f = _p.sub(p.mul(power_basis));
-		uint256 d1 = _d;
-		uint256 d2 = _d;
-		for (uint256 i = 0; i < p; i++) {
-			d1 = (d1.mul(_d)).div(basis);
-		}
-		for (uint256 i = 0; i < rp; i++) {
-			d2 = (d2.mul(_d)).div(basis);
-		}
-		uint256 g = ((d1.sub(d2)).mul(f)).div(power_basis);
-		uint256 d = d1.sub(g);
-		uint256 mint = max.mul(d);
-		mint = mint.div(basis);
-		return mint;
+		uint256 totalSupply = IERC20(registry().registries("Dev"))
+			.totalSupply();
+		return
+			curveRewards(
+				_lockups,
+				_assets,
+				totalSupply,
+				mint_per_block_and_aseet
+			);
 	}
 
 	function holdersShare(uint256 _reward, uint256 _lockups)
@@ -79,21 +66,28 @@ contract DIP1 is IPolicy, UsingRegistry {
 		return a.sub(b);
 	}
 
-	function shareOfTreasury(uint256)
+	function shareOfTreasury(uint256 _supply)
 		external
-		view
-		virtual
+		pure
 		override
 		returns (uint256)
 	{
-		return 0;
+		return _supply.div(100).mul(5);
 	}
 
-	function treasury() external view virtual override returns (address) {
-		return address(0);
+	function treasury() external view override returns (address) {
+		return treasuryAddress;
+	}
+
+	function setTreasury(address _treasury) external onlyOwner {
+		treasuryAddress = _treasury;
 	}
 
 	function capSetter() external view virtual override returns (address) {
-		return address(0);
+		return capSetterAddress;
+	}
+
+	function setCapSetter(address _setter) external onlyOwner {
+		capSetterAddress = _setter;
 	}
 }
